FEAT #1 Positioned thruster in local coordinates
Thought I had already done this, but the coordinates I passed never made it into Box2D for consideration. So now a drive can accept config.x, config.y, and config.angle, which will be resolved in local coordinate space when the force is applied.

Hooray for body.GetLocalPoint() doing the coordinate math for me!