FEAT #5, #7, and #8 Add satellite model
Figured out adding multiple fixtures to a body to create complex concave polygons. Using Crafty.viewport.mouselook(true) to enable click-to-drag viewport movement. Unfortunately, Box2D is not aware of this viewport movement, so in order to really playtest as we develop, we may have to start throwing in some graphics earlier than intended, just to see what we're doing. For now, I've added the walls back in to constrain movement to the bounds of the stage. Maybe that will hold us over until we're really ready to work on graphics and the polygon shapes underpinning them in Box2D.