FEAT #10 Dude got arms
Astronaut now has two arms, dynamic bodies with revolute joints to the shoulder. First attempt produced bizarre phantom rotational forces, which I finally diagnosed as the arms colliding (overlapping, really) with the EVA pack. I knew they wouldn't interfere with the astronaut's body, but the EVA pack is a separate physics object. So both arms, the body, and the EVA are all now part of collision group index -1. Negatives never collide with each other, positives always do, regardless of other category and mask settings. I think there are only 16 groups available (-8..8). For this game that's unlikely to pose a problem, but it's good to note that for more complex interactions, we may need to juggle those groups on the fly.