FEAT #12 Grapples on to debris, but can't let go!
Friction joint was actually pretty simple to implement, but maybe that was just thanks to recent experience doing revolute joints for the shoulders. Funny thing about friction joints is that the max force and torque settings only apply to slippage; they don't have any awareness of when the joint should outright break. So, the astronaut will slip and slide quite nicely until he is three meters away from the object he's "holding on to" and still has a firm joint applied. Obviously, that's the next thing to tackle.