FEAT #10 Catch and release
Astronaut now has a grasping property and a letGo() method. The contact listener will only create a new friction joint at the astronaut's hand if astronaut.grasping is true. The letGo() method removes friction joints from both hands, if any, and toggles grasping off for one second. Additionally, if the distance between the friction joint anchor points exceeds 15cm, the joint is destroyed. The astronaut.grasping property is unchanged in such an involuntary situation, so the astronaut may be able to slide or rotate to gain a new grip, or even grab a different object.